% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguen_grafik.R
\name{penguen_grafik}
\alias{penguen_grafik}
\title{Penguen Gaga Uzunluğu Grafiği}
\usage{
penguen_grafik(tur = NULL)
}
\arguments{
\item{tur}{Karakter. Tür filtresi (opsiyonel).}
}
\value{
ggplot nesnesi
}
\description{
Penguen türlerine göre gaga uzunluğu dağılımını gösterir.
}
