% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temiz_penguinler}
\alias{temiz_penguinler}
\title{Temiz Penguenler Veri Seti}
\format{
A data frame with 333 rows and 6 variables:
\describe{
\item{species}{Penguen türü (Adelie, Chinstrap, Gentoo)}
\item{island}{Penguenlerin yaşadığı ada}
\item{bill_length_mm}{Gaga uzunluğu (mm cinsinden)}
\item{bill_depth_mm}{Gaga derinliği (mm cinsinden)}
\item{flipper_length_mm}{Yüzgeç uzunluğu (mm)}
\item{body_mass_g}{Vücut ağırlığı (gram)}
\item{sex}{Cinsiyet (male/female)}
\item{year}{Veri Kayıt Yılı}
}
}
\source{
palmerpenguins
}
\usage{
temiz_penguinler
}
\description{
Eksik değerleri temizlenmiş penguins veri setidir.
}
\keyword{datasets}
