% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temiz_penguinler}
\alias{temiz_penguinler}
\title{Temiz Penguinler Veri Seti}
\format{
data.frame
}
\source{
palmerpenguins
}
\usage{
temiz_penguinler
}
\description{
Eksik değerleri temizlenmiş penguins veri setidir.
}
\keyword{datasets}
